back-end/requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
Django==4.2.8
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djangorestframework-xml==2.0.0
Faker==22.0.0
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2024.1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
validate-docbr==1.10.0
